requirements.txt
flask==3.0.0
flask-cors==4.0.0
surya-ocr==0.4.0
pillow>=10.2.0
torch>=2.0.0
